I made a shopping cart page, calculated the order amount, deleted and added to the cart. Processing the order entity

+++ src/main/java/org/example/repos/ProductRepo.java
@@ -2,6 +2,6 @@ package org.example.repos;
 import org.example.domain.Product;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepo extends CrudRepository<Product, Integer> {
+public interface ProductRepo extends CrudRepository<Product, Long> {
 }
